+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog rules:
 -->
 
 ## Unreleased
+- [**BC**] Change `Stringify::stringify` second parameter from `fullOutput` to more specific `shrinkOutput` and the default value from `false` to `true` to remain functionality of `stringify` function
 
 ## 1.1.0 - 2019-04-15
 - Stringify object more verbosely - depending on object type